Extracted class NBADatabaseConnector from Start
